requirements.txt
google-api-python-client==2.91.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0